refactor: refactor data classes
