+++ app/src/main/java/org/systers/mentorship/models/User.kt
@@ -1,12 +1,16 @@
-package org.systers.mentorship.remote.requests
+package org.systers.mentorship.models
 
+import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
 
 /**
- * This data class represents all the information that can be updated by a user of the system
+ * This data class represents all the information related to a user of the system
  *
- * @param name name of the user
+ * @param id identifier of the user
  * @param username username of the user
+ * @param name name of the user
+ * @param email email of the user
  * @param bio bio of the user
  * @param location location, e.g., country or/and city where the user lives
  * @param occupation current occupation of the user
@@ -17,9 +21,12 @@ import com.google.gson.annotations.SerializedName
  * @param isAvailableToMentor true, if user is available to mentor, false if otherwise
  * @param slackUsername Slack username
  */
-data class UpdateUserRequest(
-        var name: String? = null,
+@Parcelize
+data class User(
+        var id: Int? = null,
         var username: String? = null,
+        var name: String? = null,
+        var email: String? = null,
         var bio: String? = null,
         var location: String? = null,
         var occupation: String? = null,
@@ -28,4 +35,5 @@ data class UpdateUserRequest(
         var skills: String? = null,
         @SerializedName("need_mentoring") var needsMentoring: Boolean? = null,
         @SerializedName("available_to_mentor") var isAvailableToMentor: Boolean? = null,
-        @SerializedName("slack_username") var slackUsername: String? = null)
+        @SerializedName("slack_username") var slackUsername: String? = null
+) : Parcelable
