+++ app/src/main/java/org/systers/mentorship/remote/requests/Login.kt
@@ -1,0 +1,7 @@
+package org.systers.mentorship.remote.requests
+
+/**
+ * Created by murad on 7/26/18.
+ */
+data class Login(var username: String,
+                 var password: String)
