--- app/src/main/java/org/systers/mentorship/remote/requests/LoginRequest.kt
@@ -1,8 +1,0 @@
-package org.systers.mentorship.remote.requests
-
-/**
- * Created by murad on 7/26/18.
- */
-data class LoginRequest(var username: String? = null,
-                        var password: String? = null
-)
