+++ app/src/main/java/org/systers/mentorship/remote/responses/AuthToken.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 /**
  * Created by murad on 7/26/18.
  */
-data class LoginResponse(@SerializedName("access_token") var authToken: String,
-                         var expiry: Float)
+data class AuthToken(@SerializedName("access_token") var authToken: String,
+                     var expiry: Float)
