+++ app/src/main/java/org/systers/mentorship/remote/responses/MentorshipRelationResponse.kt
@@ -1,8 +1,8 @@
 package org.systers.mentorship.remote.responses
 
-import android.os.Parcel
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
 
 /**
  * This data class represents partial information of user of the system.
@@ -21,6 +21,7 @@ import com.google.gson.annotations.SerializedName
  * @param state state of the relation (@link to MentorshipRelationState)
  * @param notes notes related to the mentorship relation
  */
+@Parcelize
 data class MentorshipRelationResponse(
         val id: Int,
         @SerializedName("action_user_id") val actionUserId: Int,
@@ -33,48 +34,4 @@ data class MentorshipRelationResponse(
         @SerializedName("end_date")  val endAtTimestamp: Float,
         val state: Int,
         val notes: String
-) : Parcelable {
-    constructor(parcel: Parcel) : this(
-            parcel.readInt(),
-            parcel.readInt(),
-            parcel.readByte() != 0.toByte(),
-            parcel.readParcelable(RelationUserResponse::class.java.classLoader),
-            parcel.readParcelable(RelationUserResponse::class.java.classLoader),
-            parcel.readFloat(),
-            parcel.readFloat(),
-            parcel.readFloat(),
-            parcel.readFloat(),
-            parcel.readInt(),
-            parcel.readString())
-
-    override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(id)
-        parcel.writeInt(actionUserId)
-        parcel.writeByte(if (sentByMe) 1 else 0)
-        parcel.writeParcelable(mentor, flags)
-        parcel.writeParcelable(mentee, flags)
-        parcel.writeFloat(createdAtTimestamp)
-        parcel.writeFloat(acceptedAtTimestamp)
-        parcel.writeFloat(startAtTimestamp)
-        parcel.writeFloat(endAtTimestamp)
-        parcel.writeInt(state)
-        parcel.writeString(notes)
-    }
-
-    override fun describeContents(): Int {
-        return 0
-    }
-
-    /**
-     * Class responsible for generating instances of this Parcelable class from a Parcel
-     */
-    companion object CREATOR : Parcelable.Creator<MentorshipRelationResponse> {
-        override fun createFromParcel(parcel: Parcel): MentorshipRelationResponse {
-            return MentorshipRelationResponse(parcel)
-        }
-
-        override fun newArray(size: Int): Array<MentorshipRelationResponse?> {
-            return arrayOfNulls(size)
-        }
-    }
-}
+): Parcelable
