+++ app/src/main/java/org/systers/mentorship/remote/responses/RelationUserResponse.kt
@@ -1,7 +1,7 @@
 package org.systers.mentorship.remote.responses
 
-import android.os.Parcel
 import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
 
 /**
  * This data class represents partial information of user of the system.
@@ -11,31 +11,6 @@ import android.os.Parcelable
  * @param id identifier of the user
  * @param name name of the user
  */
-data class RelationUserResponse (val id: Int, val name: String) : Parcelable {
-
-    constructor(parcel: Parcel) : this(
-            parcel.readInt(),
-            parcel.readString())
-
-    override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(id)
-        parcel.writeString(name)
-    }
-
-    override fun describeContents(): Int {
-        return 0
-    }
-    
-    /**
-     * Class responsible for generating instances of this Parcelable class from a Parcel
-     */
-    companion object CREATOR : Parcelable.Creator<RelationUserResponse> {
-        override fun createFromParcel(parcel: Parcel): RelationUserResponse {
-            return RelationUserResponse(parcel)
-        }
-
-        override fun newArray(size: Int): Array<RelationUserResponse?> {
-            return arrayOfNulls(size)
-        }
-    }
-}
+@Parcelize
+data class RelationUserResponse (val id: Int,
+                                 val name: String) : Parcelable
